Move each cryo tank to be one tech node above corresponding "plain" tank
Each cryo tank (well, except the huge 10m spherical one) is similar in
size and volume to one of the stock non-cryo tank.  This change moves
the cryo tanks in the tech tree so that each is one step above the
corresponding non-cryo tank, to represent the extra tech involved in the
cryo tanks' "insulation and advanced materials".  The huge 10m tank is a
step above everything else because, well, it's huge.

In particular, this change moves all the tanks out of the "nuclear
propulsion" node and into lower nodes, which is appropriate since
they're useful with non-nuclear cryo engines in those nodes.